refactor(arch): 砍断 5 个 server 反向依赖环（搬符号下沉 + lazy→eager）
按消环 spec 把被 server 反向依赖的符号下沉到双方都能依赖的下层，消掉 5 个环：
- tools→server：trace 下沉 telemetry/，删死代码 os_message，解散 tools/（git_utils/
  workspace/deployment_agent 移入 cli/）
- session→server：config_service 下沉新建 config/ 叶子包
- recipes→server：github_rate_limit 移入 recipes/
- server↔agent_driver（纯 lazy 双向环）：transcript_completion + claude_sessions 下沉 session/
- viewer→server：规则层承认为合法正向特性依赖（代码零改动）

消环后把续命的 lazy import 还原成顶层 eager，作为「环已消」硬证据。行为冻结、纯 git mv。
lint-imports broken 4→2（剩 chrome-extension + recipes→chrome open_url 永久例外）。
