+++ src/frago/config/__init__.py
@@ -1,0 +1,1 @@
+"""frago config — pure leaf package for configuration services."""
